restricted pages

+++ python_learning/base/views.py
@@ -1,7 +1,9 @@
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db.models import Q
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 
 from .forms import RoomForm
@@ -9,6 +11,9 @@ from .models import Room, Topic
 
 
 def login_user(request):
+    if request.user.is_authenticated:
+        return redirect("home")
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -53,6 +58,7 @@ def room(request, pk):
     return render(request, "room.html", context)
 
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -65,9 +71,13 @@ def create_room(request):
     return render(request, "room_form.html", context)
 
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You don`t have required permissions")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -80,8 +90,12 @@ def update_room(request, pk):
     return render(request, "room_form.html", context)
 
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You don`t have required permissions")
 
     if request.method == "POST":
         room.delete()
